Revert "moved components folder inside app directory"
This reverts commit c7ccc06d82235806b2584918e5aa4d9840de9dab.
